% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Cajanuma.R
\docType{data}
\name{Cajanuma}
\alias{Cajanuma}
\title{Cajanuma}
\format{
A List of length 6 .
}
\usage{
Cajanuma
}
\description{
A list containing the data form a hummingbird-plant community in Southern Ecuador.
The Datq consist of an interaction network along with infromation on the species' morphologies, abundances, and phenologies.
}
\keyword{datasets}
